Proposition pour les money filter (#829)
* proposition pour les money filter

* fix unit test
