+++ src/filters/filters-plugin.ts
@@ -1,8 +1,8 @@
 import Vue, { PluginObject } from 'vue';
 import LoggerPlugin from '../utils/logger/logger';
 import DateFilterPlugin from './date/date';
-import MoneyPlugin from './date/money/money';
 import I18nFilterPlugin from './i18n/i18n';
+import MoneyPlugin from './money/money';
 
 const FiltersPlugin: PluginObject<any> = {
     install(v, options): void {
